docker-swarm/microservices-demo/api/app/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
pydantic==1.10.2
psycopg2==2.9.9
requests==2.31.0
pandas==2.1.2
redis==5.0.1

opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-psycopg2==0.51b0 